// example-typescript-axios-mock/src/models/IUserModel.ts
interface IUserModel {
  id: string;
  username: string;
  password: string;
  dni: string;
  name: string;
  fatherLastName: string;
  motherLastName: string;
  birthDate: string;
  telephone: string;
  address: string;
  email: string;
  type: string;
  creationDate: string;
  updateDate: string;
  createdBy: string;
  updatedBy: string;
}

export default IUserModel;